Register VM: fix add/sub instruction with signed argument

+++ NULLC/Executor_RegVm.cpp
@@ -773,7 +773,7 @@ RegVmReturnType ExecutorRegVm::RunCode(RegVmCmd *instruction, RegVmRegister * co
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrInt));
 			REGVM_DEBUG(regFilePtr[cmd.rA].activeType = rvrInt);
 
-			regFilePtr[cmd.rA].intValue = regFilePtr[cmd.rB].intValue + cmd.argument;
+			regFilePtr[cmd.rA].intValue = regFilePtr[cmd.rB].intValue + (int)cmd.argument;
 			break;
 		case rviSub:
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrInt));
@@ -786,7 +786,7 @@ RegVmReturnType ExecutorRegVm::RunCode(RegVmCmd *instruction, RegVmRegister * co
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrInt));
 			REGVM_DEBUG(regFilePtr[cmd.rA].activeType = rvrInt);
 
-			regFilePtr[cmd.rA].intValue = regFilePtr[cmd.rB].intValue - cmd.argument;
+			regFilePtr[cmd.rA].intValue = regFilePtr[cmd.rB].intValue - (int)cmd.argument;
 			break;
 		case rviMul:
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrInt));
@@ -919,7 +919,7 @@ RegVmReturnType ExecutorRegVm::RunCode(RegVmCmd *instruction, RegVmRegister * co
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrLong));
 			REGVM_DEBUG(regFilePtr[cmd.rA].activeType = rvrLong);
 
-			regFilePtr[cmd.rA].longValue = regFilePtr[cmd.rB].longValue + cmd.argument;
+			regFilePtr[cmd.rA].longValue = regFilePtr[cmd.rB].longValue + (int)cmd.argument;
 			break;
 		case rviSubl:
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrLong));
@@ -932,7 +932,7 @@ RegVmReturnType ExecutorRegVm::RunCode(RegVmCmd *instruction, RegVmRegister * co
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrLong));
 			REGVM_DEBUG(regFilePtr[cmd.rA].activeType = rvrLong);
 
-			regFilePtr[cmd.rA].longValue = regFilePtr[cmd.rB].longValue - cmd.argument;
+			regFilePtr[cmd.rA].longValue = regFilePtr[cmd.rB].longValue - (int)cmd.argument;
 			break;
 		case rviMull:
 			REGVM_DEBUG(assert(regFilePtr[cmd.rB].activeType == rvrLong));
